tests

+++ tests/valist/valist-float-dll.c
@@ -6,7 +6,7 @@ use_va_args (const char *format, ...)
   va_list argv;
   va_start (argv, format);
   int a = va_arg (argv, int);
-  int b = va_arg (argv, int);
+  double b = va_arg (argv, double);
   va_end (argv);
   return a + b;
 }
